requirements.txt
mediapipe
opencv-python
numpy
tensorflow
scikit-learn
keyboard
mouse
PyYAML
